feat : 메인 네비게이션 컴포넌트 설정
