+++ app/src/main/java/com/viewpoint/dangder/view/chat/ChatListFragment.kt
@@ -1,22 +1,23 @@
-package com.viewpoint.dangder.view.today
+package com.viewpoint.dangder.view.chat
 
 import com.viewpoint.dangder.R
 import com.viewpoint.dangder.base.BaseFragment
+import com.viewpoint.dangder.databinding.FragmentChatListBinding
 import com.viewpoint.dangder.databinding.FragmentTodayBinding
 
-class TodayFragment : BaseFragment<FragmentTodayBinding>() {
+class ChatListFragment : BaseFragment<FragmentChatListBinding>() {
     override val layoutId: Int
-        get() = R.layout.fragment_today
+        get() = R.layout.fragment_chat_list
 
     override fun initView() {
-        TODO("Not yet implemented")
+
     }
 
     override fun subscribeModel() {
-        TODO("Not yet implemented")
+
     }
 
     override fun initData() {
-        TODO("Not yet implemented")
+
     }
 }
