+++ app/src/main/java/com/viewpoint/dangder/view/chat/ChatRoomFragment.kt
@@ -3,21 +3,22 @@ package com.viewpoint.dangder.view.chat
 import com.viewpoint.dangder.R
 import com.viewpoint.dangder.base.BaseFragment
 import com.viewpoint.dangder.databinding.FragmentChatListBinding
+import com.viewpoint.dangder.databinding.FragmentChatRoomBinding
 import com.viewpoint.dangder.databinding.FragmentTodayBinding
 
-class ChatListFragment : BaseFragment<FragmentChatListBinding>() {
+class ChatRoomFragment : BaseFragment<FragmentChatRoomBinding>() {
     override val layoutId: Int
-        get() = R.layout.fragment_chat_list
+        get() = R.layout.fragment_chat_room
 
     override fun initView() {
-        TODO("Not yet implemented")
+
     }
 
     override fun subscribeModel() {
-        TODO("Not yet implemented")
+
     }
 
     override fun initData() {
-        TODO("Not yet implemented")
+
     }
 }
