+++ app/src/main/java/com/viewpoint/dangder/view/myprofile/MyProfileFragment.kt
@@ -1,23 +1,24 @@
-package com.viewpoint.dangder.view.chat
+package com.viewpoint.dangder.view.myprofile
 
 import com.viewpoint.dangder.R
 import com.viewpoint.dangder.base.BaseFragment
 import com.viewpoint.dangder.databinding.FragmentChatListBinding
+import com.viewpoint.dangder.databinding.FragmentMyProfileBinding
 import com.viewpoint.dangder.databinding.FragmentTodayBinding
 
-class ChatRoomFragment : BaseFragment<FragmentChatListBinding>() {
+class MyProfileFragment : BaseFragment<FragmentMyProfileBinding>() {
     override val layoutId: Int
-        get() = R.layout.fragment_chat_room
+        get() = R.layout.fragment_my_profile
 
     override fun initView() {
-        TODO("Not yet implemented")
+
     }
 
     override fun subscribeModel() {
-        TODO("Not yet implemented")
+
     }
 
     override fun initData() {
-        TODO("Not yet implemented")
+
     }
 }
