+++ app/src/main/java/com/viewpoint/dangder/view/today/TodayFragment.kt
@@ -1,4 +1,22 @@
 package com.viewpoint.dangder.view.today
 
-class TodayFragment {
+import com.viewpoint.dangder.R
+import com.viewpoint.dangder.base.BaseFragment
+import com.viewpoint.dangder.databinding.FragmentTodayBinding
+
+class TodayFragment : BaseFragment<FragmentTodayBinding>() {
+    override val layoutId: Int
+        get() = R.layout.fragment_today
+
+    override fun initView() {
+
+    }
+
+    override fun subscribeModel() {
+
+    }
+
+    override fun initData() {
+
+    }
 }
