+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -1,28 +1,31 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/signip_nav_graph"
-    app:startDestination="@id/userEmailFragment">
+    android:id="@+id/main_nav_graph"
+    app:startDestination="@id/mainFragment">
+
     <fragment
-        android:id="@+id/userEmailFragment"
-        android:name="com.viewpoint.dangder.view.signup.UserEmailFragment"
-        android:label="회원가입" >
+        android:id="@+id/mainFragment"
+        android:name="com.viewpoint.dangder.view.main.MainFragment"
+        android:label="MainFragment" />
+    <fragment
+        android:id="@+id/todayFragment"
+        android:name="com.viewpoint.dangder.view.today.TodayFragment"
+        android:label="TodayFragment" />
+    <fragment
+        android:id="@+id/chatListFragment"
+        android:name="com.viewpoint.dangder.view.chat.ChatListFragment"
+        android:label="ChatListFragment" >
         <action
-            android:id="@+id/action_userEmailFragment_to_emailVerifyFragment"
-            app:destination="@id/emailVerifyFragment"
-           />
+            android:id="@+id/action_chatListFragment_to_chatRoomFragment"
+            app:destination="@id/chatRoomFragment" />
     </fragment>
     <fragment
-        android:id="@+id/emailVerifyFragment"
-        android:name="com.viewpoint.dangder.view.signup.VerifyEmailTokenFragment"
-        android:label="회원가입" >
-        <action
-            android:id="@+id/action_emailVerifyFragment_to_userPasswordFragment"
-            app:destination="@id/userPasswordFragment"
-          />
-    </fragment>
+        android:id="@+id/myProfileFragment"
+        android:name="com.viewpoint.dangder.view.myprofile.MyProfileFragment"
+        android:label="MyProfileFragment" />
     <fragment
-        android:id="@+id/userPasswordFragment"
-        android:name="com.viewpoint.dangder.view.signup.UserPasswordFragment"
-        android:label="회원가입" />
+        android:id="@+id/chatRoomFragment"
+        android:name="com.viewpoint.dangder.view.chat.ChatRoomFragment"
+        android:label="ChatRoomFragment" />
 </navigation>
